+++ src/main/java/com/Transaction/transaction/exception/ResourceNotFoundException.java
@@ -8,9 +8,9 @@ import lombok.Setter;
 public class ResourceNotFoundException extends RuntimeException {
     private String resourceName;
     private String fieldName;
-    private int fieldValue;
+    private long fieldValue;
 
-    public ResourceNotFoundException(String resourceName, String fieldName, int fieldValue) {
+    public ResourceNotFoundException(String resourceName, String fieldName, long fieldValue) {
         super(String.format("%s NOT FOUND WITH %s:%s", resourceName, fieldName, fieldValue));
         this.resourceName = resourceName;
         this.fieldName = fieldName;
