+++ src/main/java/com/Transaction/transaction/repository/BookingRepo.java
@@ -3,6 +3,6 @@ package com.Transaction.transaction.repository;
 import com.Transaction.transaction.entity.BookingTicket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookingRepo extends JpaRepository<BookingTicket, Integer> {
+public interface BookingRepo extends JpaRepository<BookingTicket, Long> {
 
 }
