+++ src/main/java/com/Transaction/transaction/repository/BusStopRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BusStopRepo extends JpaRepository<BusStop, Integer> {
+public interface BusStopRepo extends JpaRepository<BusStop, Long> {
 
 }
