+++ src/main/java/com/Transaction/transaction/repository/RouteRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RouteRepo extends JpaRepository<Route, Integer> {
+public interface RouteRepo extends JpaRepository<Route, Long> {
     List<Route> findByDestinationBusStop(Route route12);
 }
