+++ src/main/java/com/Transaction/transaction/repository/TicketRepo.java
@@ -3,5 +3,5 @@ package com.Transaction.transaction.repository;
 import com.Transaction.transaction.entity.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketRepo extends JpaRepository<Ticket, Integer> {
+public interface TicketRepo extends JpaRepository<Ticket, Long> {
 }
