+++ src/main/java/com/Transaction/transaction/service/BookingTicketService.java
@@ -7,10 +7,10 @@ import java.util.List;
 public interface BookingTicketService {
     List<BookingTicketDto> getAllBooking();
 
-    BookingTicketDto getBooking(int bookingId);
+    BookingTicketDto getBooking(long bookingId);
 
     BookingTicketDto createBooking(BookingTicketDto bookingTicketDto);
 
-    BookingTicketDto getBookById(int id);
+    BookingTicketDto getBookById(long id);
 
 }
