+++ src/main/java/com/Transaction/transaction/service/BusService.java
@@ -6,11 +6,11 @@ import java.time.LocalDate;
 import java.util.List;
 
 public interface BusService {
-    BusDto updateBusInfo(BusDto busDto, int id, int routeId);
+    BusDto updateBusInfo(BusDto busDto, long id, long routeId);
 
-    void deleteBusInfo(Integer id);
+    void deleteBusInfo(long id);
 
-    BusDto createBusForRoute(BusDto busDto, int id);
+    BusDto createBusForRoute(BusDto busDto, long id);
 
     List<BusDto> getAllBusInfo();
