+++ src/main/java/com/Transaction/transaction/service/SeatService.java
@@ -6,15 +6,15 @@ import com.Transaction.transaction.payloads.SeatDto;
 import java.util.List;
 
 public interface SeatService {
-    SeatDto updateSeat(SeatDto seatDto, int id);
+    SeatDto updateSeat(SeatDto seatDto, long id);
 
-    void deleteSeat(int id);
+    void deleteSeat(long id);
 
-    SeatDto getSeatById(int id);
+    SeatDto getSeatById(long id);
 
     List<SeatDto> getAllSeat();
 
-    SeatDto createSeatForBus(SeatDto seatDto, int id);
+    SeatDto createSeatForBus(SeatDto seatDto, long id);
 
     List<SeatDto> findSeatRelatedToBus(String busName);
